modernize-avoid-c-style-cast
Squashed commit of the following:

commit 04eb386a0c2c2337d45761917c2fa85477f4201a
Author: Frederik Hennig <frederik.hennig@fau.de>
Date:   Tue Jun 23 13:50:12 2026 +0000

    fix compiler errors

commit 86d03cb1629a9f25f5c94cc2f42272ca1114d1a7
Author: Frederik Hennig <frederik.hennig@fau.de>
Date:   Tue Jun 23 10:55:17 2026 +0000

    fix more pointer casts

commit 43fbdb121712a3784729498c87463b0740f23fd2
Author: Frederik Hennig <frederik.hennig@fau.de>
Date:   Tue Jun 23 10:16:38 2026 +0000

    fix metis options static cast

commit 7e8158d4214752e77e534062c638cd5b33f0f853
Author: Frederik Hennig <frederik.hennig@fau.de>
Date:   Tue Jun 23 09:59:10 2026 +0000

    clean remaining c-style casts

commit 9b56588e649ac4b9b3bc2818e1be29dc9ab22a7b
Author: Frederik Hennig <frederik.hennig@fau.de>
Date:   Tue Jun 23 09:44:06 2026 +0000

    clean up mesh module

commit fa06a49c9c91afc8064e23f07aa3d89756f70a1e
Author: Frederik Hennig <frederik.hennig@fau.de>
Date:   Tue Jun 23 09:41:01 2026 +0000

    clean gpu module of c-style casts

commit 6cc0afdc08a6d239d3c820eee35fe2b095f28038
Author: Frederik Hennig <frederik.hennig@fau.de>
Date:   Tue Jun 23 08:44:15 2026 +0000

    fix fieldcpy const cast

commit de7d2e683691748b66bd5b6830460bf65b8ea2bf
Author: Frederik Hennig <frederik.hennig@fau.de>
Date:   Thu Jun 18 17:33:29 2026 +0200

    static_cat? meow :3

commit 5779e0e8ad4b584637f2981500808e8107c03fd7
Author: Frederik Hennig <frederik.hennig@fau.de>
Date:   Thu Jun 18 17:23:10 2026 +0200

    fixed broken std::function type

commit 5cd47291b4197d5571fbd7b4c2e07a738500eb09
Author: Frederik Hennig <frederik.hennig@fau.de>
Date:   Thu Jun 18 16:38:01 2026 +0200

    clean up geometry and gpu

commit 6d9e715cf5bdedac3ea8b582c46c4edc653ed8f7
Author: Frederik Hennig <frederik.hennig@fau.de>
Date:   Thu Jun 18 15:13:37 2026 +0200

    update mesapd_coupling

commit bd2997c807f036e2bd385ae466136c35a3036720
Author: Frederik Hennig <frederik.hennig@fau.de>
Date:   Thu Jun 18 14:56:24 2026 +0200

    many more casts eliminated

commit 0416d160b17a36f23a5733f9c3707ba0e91fcc82
Author: Frederik Hennig <frederik.hennig@fau.de>
Date:   Thu Jun 18 14:32:33 2026 +0200

    fix all c-style-casts in field module

commit 524ecc5e211aed0240efa9c93f53fda9bc58cf01
Author: Frederik Hennig <frederik.hennig@fau.de>
Date:   Thu Jun 18 14:13:43 2026 +0200

    fix clang-format action on cuda kernel calls

commit 787cda156c4f157d4f592fae9515abcbfc6f90e4
Author: Frederik Hennig <frederik.hennig@fau.de>
Date:   Wed Jun 17 08:07:57 2026 +0200

    fix copy-paste error in TestBufferSystem

commit 32bf4fe667dfd40cd01863671b914a4af2057927
Author: Frederik Hennig <frederik.hennig@fau.de>
Date:   Tue Jun 16 17:43:09 2026 +0200

    fix invalid replacement in jinja_filters

commit 81401d4c661f92491c0d07b3086091478dcd1677
Author: Frederik Hennig <frederik.hennig@fau.de>
Date:   Tue Jun 16 15:51:02 2026 +0200

    fix even even even more constants. complete blockforest module

commit 45931da24a5e1969c1428d835b404f70770fc0b4
Author: Frederik Hennig <frederik.hennig@fau.de>
Date:   Tue Jun 16 13:44:03 2026 +0200

    one last static_cast in core

commit b3001a14f41ec7d32121bdfe68177d434411a3d7
Author: Frederik Hennig <frederik.hennig@fau.de>
Date:   Tue Jun 16 13:43:15 2026 +0200

    more real_t constants

commit 597f582e4215b648ca08d199a205e77bcdabbee3
Author: Frederik Hennig <frederik.hennig@fau.de>
Date:   Tue Jun 16 13:40:31 2026 +0200

    real_t constants

commit 98cdbc2cdf5b26cc0930d6a686ea1f7686dd801d
Author: Frederik Hennig <frederik.hennig@fau.de>
Date:   Tue Jun 16 13:39:38 2026 +0200

    mpirank casts

commit a3dac2c334019f11de06d533e1cc3466d5667903
Author: Frederik Hennig <frederik.hennig@fau.de>
Date:   Tue Jun 16 13:36:59 2026 +0200

    uint_t constants

commit 113707271ee188bf70862436401f53c45b701795
Author: Frederik Hennig <frederik.hennig@fau.de>
Date:   Tue Jun 16 13:30:08 2026 +0200

    many other uint_t and real_t constants

commit d6ef89e68b45d516601e20e55e59879795d67209
Author: Frederik Hennig <frederik.hennig@fau.de>
Date:   Tue Jun 16 13:15:51 2026 +0200

    fix real_t 0 and 1 constants

commit 5d5567651027ed2e23b42e2e60ed9755a835e3cf
Author: Frederik Hennig <frederik.hennig@fau.de>
Date:   Tue Jun 16 13:13:46 2026 +0200

    fix more constants

commit f953561f5b39b5a28296b4dfa752b97c918d8155
Author: Frederik Hennig <frederik.hennig@fau.de>
Date:   Tue Jun 16 13:10:49 2026 +0200

    fix 0 and 1 size_t constants

commit 67dcd7f4c5d3080f49d7965145156cee802814e0
Author: Frederik Hennig <frederik.hennig@fau.de>
Date:   Tue Jun 16 13:10:09 2026 +0200

    fix 0 and 1 uint_t constants

commit fe33b9539a251c83e94fc6d714379a0cc5c7fbf0
Author: Frederik Hennig <frederik.hennig@fau.de>
Date:   Tue Jun 16 13:09:05 2026 +0200

    fix tolower

commit 197f03bcdf16d25461d8d919550923e3e74c9dd1
Author: Frederik Hennig <frederik.hennig@fau.de>
Date:   Tue Jun 16 12:42:58 2026 +0200

    fix c-style-casts in core WIP

commit c2e3be2e90185be43f32c49f8da13d10d20ac69b
Author: Frederik Hennig <frederik.hennig@fau.de>
Date:   Tue Jun 16 12:31:55 2026 +0200

    fix avoid-c-style-casts in walberla::v8

commit 9f66b84afcc0d5f5fbcf76904da868eb1b80acfb
Author: Frederik Hennig <frederik.hennig@fau.de>
Date:   Tue Jun 16 12:02:38 2026 +0200

    start replacing c-style casts

+++ src/core/math/FastInvSqrt.h
@@ -43,7 +43,7 @@ namespace math {
    template<unsigned int numIter>
    inline double fastInvSqrt( double y )
    {
-      double yhalf = ( double )0.5 * y;
+      double yhalf = 0.5 * y;
       DoubleAndLongLong u;
       u.asReal = y;
       u.asInt = 0x5fe6ec85e7de30daLL - ( u.asInt >> 1 );
